\name{emergencystart}
\alias{emergencystart}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Reset starting values to simple defaults
}
\description{
This function overwrites the starting values to simple defaults. This can help in cases where optimization fails.
}
\usage{
emergencystart(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A \code{psychonetrics} model.
}
}

\value{
A \code{psychonetrics} model.
}

\author{
Sacha Epskamp
}

