\name{fit}
\alias{fit}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print fit indices
}
\description{
This function will print all fit indices of the model/
}
\usage{
fit(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A \code{psychonetrics} model.
}
}
\author{
Sacha Epskamp
}
