\name{Jonas}
\alias{Jonas}
\docType{data}
\title{
Jonas dataset
}
\description{
Responses of 10 attitude items towards a researcher named Jonas. Participants were shown three photos of Jonas with the text: "This is Jonas, a researcher from Germany who is now becoming a PhD in Psychology". Subsequently, the participants had to answer 10 yes / no questions starting with "I believe that Jonas...", as well as rate their familliarity with Jonas. The sample consists of people familiar with Jonas and not familiar with Jonas, and allows for testing Attitudinal Entropy Framework <doi:10.1080/1047840X.2018.1537246>. 
}
\usage{data("Jonas")}
\format{
  A data frame with 215 observations on the following 12 variables. 
  \describe{
    \item{\code{scientist}}{... is a good scientist}
    \item{\code{jeans}}{... Is a person that wears beautiful jeans}
    \item{\code{cares}}{... really cares about people like you}
    \item{\code{economics}}{... would solve our economic problems}
    \item{\code{hardworking}}{... is hardworking}
    \item{\code{honest}}{... is honest}
    \item{\code{intouch}}{... is in touch with ordinary people}
    \item{\code{knowledgeable}}{... is knowledgeable}
    \item{\code{makeupmind}}{... can't make up his mind}
    \item{\code{getsthingsdone}}{... gets things done}
    \item{\code{familiar}}{Answers to the question "How familiar are you with Jonas?" (three responses possible)}
    \item{\code{group}}{The question 'familiar' categorized in two groups ("Knows Jonas" and "Doesn't Know Jonas")}
  }
}

\examples{
data(Jonas)
}
\keyword{datasets}
