\name{logbook}
\alias{logbook}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Retrieve the psychonetrics logbook
}
\description{
This function can be used to retrieve the logbook of a 'psychonetrics' object.
}
\usage{
logbook(x, log = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
 A 'psychonetrics' object.
 }
  \item{log}{
Logical, should the entry that the logbook is accessed be added?
}
}

\author{
Sacha Epskamp
}

