\name{psychonetrics-package}
\alias{psychonetrics-package}
\alias{psychonetrics}
\docType{package}
\title{
\packageTitle{psychonetrics}
}
\description{
\packageDescription{psychonetrics}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{psychonetrics}
\packageIndices{psychonetrics}
This package can be used to perform Structural Equation Modeling and confirmatory network modeling. Current implemented families of models are (1) the variance--covariance matrix (\code{\link{varcov}}), (2) the latent variable model (\code{\link{lvm}}), (3) the lag-1 vector autoregression model (\code{\link{var1}}), (4) the dynamical lag-1 latent variable model for panel data (\code{\link{dlvm1}}) and for time-series data (\code{\link{tsdlvm1}}), (5) the Ising model (\code{\link{Ising}}) and its generalization the Blume-Capel model (\code{\link{BlumeCapel}}), (6) the multi-level latent variable model for two-level data (\code{\link{ml_lvm}}) and the multi-level dynamical model for multi-subject time-series data (\code{\link{ml_tsdlvm1}}), (7) the meta-analytic variance--covariance model (\code{\link{meta_varcov}}, with wrappers \code{\link{meta_lvm}} and \code{\link{meta_var1}}), and (8) the random-intercept cross-lagged panel model (\code{\link{ri_clpm}}).
}
\author{
\packageAuthor{psychonetrics}

Maintainer: \packageMaintainer{psychonetrics}
}
\references{
More information: psychonetrics.org
}
