\name{setverbose}
\alias{setverbose}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Should messages of computation progress be printed?
}
\description{
This function controls if messages should be printed for a psychonetrics model. 
}
\usage{
setverbose(x, verbose = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A \code{psychonetrics} model.
}
  \item{verbose}{
Logical indicating if verbose should be enabled
}
}
\value{
An object of the class psychonetrics (\link{psychonetrics-class})
}
\author{
Sacha Epskamp
}