\name{getVCOV}
\alias{getVCOV}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Obtain the asymptotic covariance matrix
}
\description{
This function can be used to obtain the estimated asymptotic covariance matrix from a \code{psychonetrics} object.
}
\usage{
getVCOV(model)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{model}{
A \code{psychonetrics} model.
}
}

\value{
This function returns a matrix.
}
\author{
Sacha Epskamp
}
