\name{parameters}
\alias{parameters}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print parameter estimates
}
\description{
This function will print a list of parameters of the model
}
\usage{
parameters(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A \code{psychonetrics} model.
}
}
\author{
Sacha Epskamp
}