\name{psychonetrics-package}
\alias{psychonetrics-package}
\alias{psychonetrics}
\docType{package}
\title{
\packageTitle{psychonetrics}
}
\description{
\packageDescription{psychonetrics}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{psychonetrics}
\packageIndices{psychonetrics}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{psychonetrics}

Maintainer: \packageMaintainer{psychonetrics}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
