\name{psychonetrics-package}
\alias{psychonetrics-package}
\alias{psychonetrics}
\docType{package}
\title{
\packageTitle{psychonetrics}
}
\description{
\packageDescription{psychonetrics}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{psychonetrics}
\packageIndices{psychonetrics}
This package can be used to perform Structural Equation Modeling and confirmatory network modeling. Current implemented families of models are (1) the variance--covariance matrix (\code{\link{varcov}}), (2) the latent variable model (\code{\link{lvm}}), (3) the lag-1 vector autoregression model (\code{\link{var1}}), and (4) the dynamical lag-1 latent variable model for panel data (\code{\link{dlvm1}}) and for time-series data (\code{\link{tsdlvm1}}).
}
\author{
\packageAuthor{psychonetrics}

Maintainer: \packageMaintainer{psychonetrics}
}
\references{
More information: psychonetrics.org
}
